requirements.txt
absl-py==2.1.0
python-docx==1.1.2
